// data/products/detail/ea-product-details.zh.generated.ts
export const eaProductDetailStaticParams = eaProductDetailsZh.map(
  (item) => ({
    category: item.category,
    slug: item.slug,
  }),
);